updating BaseAPIHandler
