+++ consoleme/handlers/v2/index.py
@@ -3,7 +3,7 @@ from operator import itemgetter
 from urllib.parse import parse_qs, urlencode, urlparse
 
 from consoleme.config import config
-from consoleme.handlers.base import BaseHandler, BaseAPIV1Handler
+from consoleme.handlers.base import BaseHandler
 from consoleme.lib.handler_utils import format_role_name
 from consoleme.lib.loader import WebpackLoader
 from consoleme.lib.plugins import get_plugin_by_name
@@ -175,17 +175,3 @@ class IndexHandler(BaseHandler):
 
         self.redirect(url)
 
-
-class SelectRolesHandler(BaseAPIV1Handler):
-    def initialize(self):
-        self.user: str = None
-        self.eligible_roles: list = []
-
-    async def get(self):
-        self.set_header("Content-Type", "application/json")
-        payload = {
-            "eligible_roles": self.eligible_roles,
-            "_xsrf": self.xsrf_token.decode("utf-8"),
-        }
-        self.write(json.dumps(payload))
-        await self.finish()
