+++ consoleme/handlers/v2/roles.py
@@ -1,7 +1,7 @@
 import ujson as json
 
 from consoleme.config import config
-from consoleme.handlers.base import BaseJSONHandler, BaseAPIV2Handler
+from consoleme.handlers.base import BaseJSONHandler, BaseAPIHandler
 from consoleme.lib.crypto import Crypto
 from consoleme.lib.plugins import get_plugin_by_name
 
@@ -11,7 +11,7 @@ crypto = Crypto()
 auth = get_plugin_by_name(config.get("plugins.auth"))()
 
 
-class RolesHandler(BaseAPIV2Handler):
+class RolesHandler(BaseAPIHandler):
     """Handler for /api/v2/roles
 
     Allows read access to a list of roles across all accounts. Returned roles are
@@ -20,26 +20,21 @@ class RolesHandler(BaseAPIV2Handler):
 
     allowed_methods = ["GET"]
 
-    def __init__(self, *args, **kwargs):
-        # TODO(psanders): Use actual JWT validator
-        super().__init__(jwt_validator=lambda x: {}, *args, **kwargs)
+    def initialize(self):
+        self.user: str = None
+        self.eligible_roles: list = []
 
     async def get(self):
-        """
-        GET /api/v2/roles
-        """
-        log_data = {
-            "function": "RolesHandler.get",
-            "user": self.user,
-            "message": "Writing all eligible user roles",
-            "user-agent": self.request.headers.get("User-Agent"),
-            "request_id": self.request_uuid,
+        self.set_header("Content-Type", "application/json")
+        payload = {
+            "eligible_roles": self.eligible_roles,
+            "_xsrf": self.xsrf_token.decode("utf-8"),
         }
-        log.debug(log_data)
-        self.write_error(501, message="Get roles")
+        self.write(json.dumps(payload, escape_forward_slashes=False))
+        await self.finish()
 
 
-class AccountRolesHandler(BaseAPIV2Handler):
+class AccountRolesHandler(BaseAPIHandler):
     """Handler for /api/v2/roles/{account_number}
 
     Allows read access to a list of roles by account. Roles are limited to what the
@@ -47,10 +42,6 @@ class AccountRolesHandler(BaseAPIV2Handler):
     """
 
     allowed_methods = ["GET"]
-
-    def __init__(self, *args, **kwargs):
-        # TODO(psanders): Use actual JWT validator
-        super().__init__(jwt_validator=lambda x: {}, *args, **kwargs)
 
     async def get(self, account_id):
         """
@@ -67,17 +58,13 @@ class AccountRolesHandler(BaseAPIV2Handler):
         self.write_error(501, message="Get roles by account")
 
 
-class RoleDetailHandler(BaseAPIV2Handler):
+class RoleDetailHandler(BaseAPIHandler):
     """Handler for /api/v2/roles/{accountNumber}/{roleName}
 
     Allows read and update access to a specific role in an account.
     """
 
     allowed_methods = ["GET", "PUT"]
-
-    def __init__(self, *args, **kwargs):
-        # TODO(psanders): Use actual JWT validator
-        super().__init__(jwt_validator=lambda x: {}, *args, **kwargs)
 
     async def get(self, account_number, role_name):
         """
